# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NexusID API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}/block:
    post:
      operationId: blockUser
      tags: [moderation]
      summary: Block a user from sending DMs
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Blocked
        "400":
          description: Cannot block yourself
        "401":
          description: Unauthorized
        "404":
          description: User not found
    delete:
      operationId: unblockUser
      tags: [moderation]
      summary: Remove an existing block on a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Unblocked
        "401":
          description: Unauthorized
  /users/{userId}/report:
    post:
      operationId: reportUser
      tags: [moderation]
      summary: File a moderation report against a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportUserBody"
      responses:
        "201":
          description: Report filed
        "400":
          description: Cannot report yourself
        "401":
          description: Unauthorized
        "404":
          description: User not found
...
  /moderation/reports:
    get:
      operationId: listUserReports
      tags: [moderation]
      summary: List user reports for moderator review (admin only)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, reviewed, dismissed, actioned, all]
            default: pending
      responses:
        "200":
          description: Reports list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserReportsList"
        "401":
          description: Unauthorized
        "403":
          description: Not an admin
  /moderation/reports/{reportId}:
    patch:
      operationId: updateUserReportStatus
      tags: [moderation]
      summary: Update the status of a user report (admin only)
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserReportBody"
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserReport"
        "400":
          description: Invalid status
        "401":
          description: Unauthorized
        "403":
          description: Not an admin
        "404":
          description: Report not found
  /moderation/reports/{reportId}/conversation:
    get:
      operationId: getUserReportConversation
      tags: [moderation]
      summary: Get the conversation snippet attached to a report (admin only)
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Conversation snippet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserReportConversation"
        "401":
          description: Unauthorized
        "403":
          description: Not an admin
        "404":
          description: Report or conversation not found
  /admin/users/{userId}/verification:
    patch:
      operationId: setUserVerificationTier
      tags: [moderation]
      summary: Grant or revoke an identity-verification badge (admin only)
      description: |
        Sets the user's verificationTier to "silver", "blue", or null
        (revoke). Surfaced as a check-mark badge next to the user's
        display name throughout the product.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetVerificationTierBody"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Invalid tier
        "401":
          description: Unauthorized
        "403":
          description: Not an admin
        "404":
          description: User not found
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [moderation]
      summary: List users for admin moderation (admin only)
      description: |
        Returns matching user profiles for the admin UI. Use `adminOnly=true`
        to fetch the current admin roster, or `q` to search by username /
        display name when granting admin to a new user.
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: adminOnly
          in: query
          schema:
            type: boolean
            default: false
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Matching users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSearchResult"
        "401":
          description: Unauthorized
        "403":
          description: Not an admin
  /admin/users/{userId}/admin:
    patch:
      operationId: setUserAdmin
      tags: [moderation]
      summary: Promote or demote a user to admin/moderator (admin only)
      description: |
        Sets the user's `isAdmin` flag. Admins cannot demote themselves —
        attempting to do so returns 400, which guards against an admin
        accidentally locking themselves out of the moderation surface.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetUserAdminBody"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Invalid input or self-demotion
        "401":
          description: Unauthorized
        "403":
          description: Not an admin
        "404":
          description: User not found

  # ─── Storage ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        occupation:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
        experience:
          type: array
          items:
            $ref: "#/components/schemas/Experience"
        didDocument:
          type: object
          nullable: true
          additionalProperties: true
        followersCount:
          type: integer
        followingCount:
          type: integer
        postsCount:
          type: integer
        ghostMode:
          type: boolean
        isFollowing:
          type: boolean
        bountiesWon:
          type: integer
        createdAt:
          type: string
        usernameChangedAt:
          type: string
          nullable: true
        lastSeenAt:
          type: string
          nullable: true
          description: Last time this user pinged the server. Powers the online dot.
        powerScore:
          type: ["integer", "null"]
        powerRank:
          type: ["string", "null"]
        isAdmin:
          type: boolean
          description: True when this user is a moderator/admin.
        verificationTier:
          type: string
          nullable: true
          enum: [silver, blue, null]
          description: |
            Identity-verification badge shown next to the display name.
            "silver" = identity verified, "blue" = notable / premium account,
            null = unverified. Granted by an admin via
            PATCH /admin/users/:userId/verification.
        chatScreenshotsTaken:
          type: integer
          description: |
            Public counter of how many times this user has screenshotted a
            DM thread (mobile only). Surfaced as a warning chip on the
            profile so peers can decide whether to share sensitive content.
      required: [id, clerkId, username, displayName, skills, experience, followersCount, followingCount, postsCount, ghostMode, isFollowing, bountiesWon, createdAt]
    SetUserAdminBody:
      type: object
      properties:
        isAdmin:
          type: boolean
          description: True to promote the user to admin, false to demote.
      required: [isAdmin]
    SetVerificationTierBody:
      type: object
      properties:
        tier:
          type: string
          nullable: true
          enum: [silver, blue, null]
          description: New verification tier. Use null to revoke the badge.
      required: [tier]
    Experience:
      type: object
      properties:
        title:
          type: string
        company:
          type: string
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        current:
          type: boolean
        description:
          type: string
          nullable: true
      required: [title, company, startDate, current]
...
    UserSearchResult:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total:
          type: integer
      required: [users, total]
...
    DirectMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: string
        recipientId:
          type: string
        content:
          type: string
        expiresAt:
          type: string
          nullable: true
        readAt:
          type: string
          nullable: true
        createdAt:
          type: string
        expired:
          type: boolean
          description: True when the message's self-destruct TTL has elapsed. The
            server scrubs `content` to an empty string in that case; the client
            should render a "self-destructed" placeholder until the cleanup
            sweep deletes the row entirely.
      required: [id, conversationId, senderId, recipientId, content, createdAt]
...
    ReportUserBody:
      type: object
      properties:
        reason:
          type: string
          maxLength: 1000
        conversationId:
          type: integer
          nullable: true
    UpdateUserReportBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, reviewed, dismissed, actioned]
      required: [status]
    UserReport:
      type: object
      properties:
        id:
          type: integer
        reporter:
          $ref: "#/components/schemas/UserProfile"
        reported:
          $ref: "#/components/schemas/UserProfile"
        conversationId:
          type: integer
          nullable: true
        reason:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, reviewed, dismissed, actioned]
        reviewedAt:
          type: string
          nullable: true
        reviewedById:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, reporter, reported, conversationId, reason, status, createdAt]
    UserReportsList:
      type: object
      properties:
        reports:
          type: array
          items:
            $ref: "#/components/schemas/UserReport"
      required: [reports]
    UserReportConversation:
      type: object
      properties:
        conversationId:
          type: integer
        messages:
          type: array
          items:
            $ref: "#/components/schemas/DirectMessage"
      required: [conversationId, messages]
...